modules/skeletor-py/requirements.txt
opencv-python
numpy
tensorflow
tensorflow_hub
Pillow
tqdm # Added for progress bar